Updated User and ChatMessage models

+++ Controllers/UserController.cs
@@ -45,7 +45,7 @@ namespace SimpleChat.Controllers
         // PUT: api/User/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut("{id}")]
-        public async Task<IActionResult> PutUser(int id, User user)
+        public async Task<IActionResult> PutUser(ulong id, User user)
         {
             if (id != user.Id)
             {
@@ -100,7 +100,7 @@ namespace SimpleChat.Controllers
             return NoContent();
         }
 
-        private bool UserExists(int id)
+        private bool UserExists(ulong id)
         {
             return _context.Users.Any(e => e.Id == id);
         }
